move the Shutdown error type to app module
